chore: generate options JSON schema automatically
