--- src/types/blaze.ts
@@ -1,7 +1,0 @@
-export type BlazeOption = {
-    name: string;
-    url: string;
-    auth?: string;
-    backendMeasures: string;
-    uiSiteMap: string[][];
-};
